+++ docs/changelog.rst
@@ -15,7 +15,7 @@ New Features
 
 Documentation
 =============
-- Adjust PDF documentation build's font size of code-blocks to fit 120 character width files (:issue:`445`). By `Kyle
+- Adjust PDF documentation build's font size of code-blocks to fit 120 character width files (:merge:`446` :merge:`446`). By `Kyle
   Brindley`_.
 
 ******************